Added support for DotNetCore
