--- Forecast.io.40.sln
@@ -1,20 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Forecast.io.40", "forecast.io\Forecast.io.40.csproj", "{505F13DA-0FF3-4B14-8A30-86D8EB95F0AC}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{505F13DA-0FF3-4B14-8A30-86D8EB95F0AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{505F13DA-0FF3-4B14-8A30-86D8EB95F0AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{505F13DA-0FF3-4B14-8A30-86D8EB95F0AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{505F13DA-0FF3-4B14-8A30-86D8EB95F0AC}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
