--- forecast.io/Entities/CompressedWebClient.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Net;
-
-namespace ForecastIO
-{
-    public class CompressionEnabledWebClient : WebClient
-    {
-        protected override WebRequest GetWebRequest(Uri address)
-        {
-            var request = base.GetWebRequest(address) as HttpWebRequest;
-            if (request == null) return null;
-            request.AutomaticDecompression = DecompressionMethods.Deflate | DecompressionMethods.GZip;
-            return request;
-        }
-        
-    }
-}
