+++ src/Forecast.io.tests/project.json
@@ -1,0 +1,29 @@
+﻿{
+  "version": "1.0.0-*",
+  "description": "Forecast.io.tests Class Library",
+  "authors": [ "fish_" ],
+  "tags": [ "" ],
+  "projectUrl": "",
+  "licenseUrl": "",
+  "dependencies": {
+    "NUnit": "3.0.1",
+    "Forecast.io": "1.0.0-*"
+  },
+  "frameworks": {
+    "net40": { },
+    "net451": { },
+    "net461": { },
+    "dnx451": { },
+    "dnx461": { },
+    "dnxcore50": {
+      "dependencies": {
+        "System.Runtime": "4.0.21-rc2-23623",
+        "System.Collections": "4.0.11-beta-23516",
+        "System.Linq": "4.0.1-beta-23516",
+        "Microsoft.CSharp": "4.0.1-beta-23516",
+        "System.Threading": "4.0.11-beta-23516",
+        "System.Console": "4.0.0-beta-23516"
+      }
+    }
+  }
+}
